requirements.txt
typer
python-moodle
python-dotenv